+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Add project layout and basic "index.html" page
 - Verify passwords based on ``wp_users`` WordPress table
+- Make ``content.html`` configurable